# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Market {
  symbol: string;
  label: string;
}

export type EngineConfigAccountType = typeof EngineConfigAccountType[keyof typeof EngineConfigAccountType];


export const EngineConfigAccountType = {
  demo: 'demo',
  real: 'real',
} as const;

export interface EngineConfig {
  /** Deriv volatility index symbol, e.g. R_100 */
  market: string;
  accountType: EngineConfigAccountType;
  /** @minimum 0.35 */
  baseStake: number;
  martingaleEnabled: boolean;
  /** @minimum 1 */
  martingaleMultiplier: number;
  /** @minimum 0 */
  stopLoss: number;
  /** @minimum 0 */
  takeProfit: number;
  /** @minimum 1 */
  maxSuccessiveLosses: number;
  /** @minimum 1 */
  maxSuccessiveWins: number;
}

export type EngineConfigUpdateAccountType = typeof EngineConfigUpdateAccountType[keyof typeof EngineConfigUpdateAccountType];


export const EngineConfigUpdateAccountType = {
  demo: 'demo',
  real: 'real',
} as const;

export interface EngineConfigUpdate {
  market?: string;
  accountType?: EngineConfigUpdateAccountType;
  /** @minimum 0.35 */
  baseStake?: number;
  martingaleEnabled?: boolean;
  /** @minimum 1 */
  martingaleMultiplier?: number;
  /** @minimum 0 */
  stopLoss?: number;
  /** @minimum 0 */
  takeProfit?: number;
  /** @minimum 1 */
  maxSuccessiveLosses?: number;
  /** @minimum 1 */
  maxSuccessiveWins?: number;
}

export type TradeRecordResult = typeof TradeRecordResult[keyof typeof TradeRecordResult];


export const TradeRecordResult = {
  win: 'win',
  loss: 'loss',
} as const;

export interface TradeRecord {
  id: string;
  timestamp: string;
  market: string;
  contractType: string;
  barrier: string;
  stake: number;
  payout: number;
  profit: number;
  result: TradeRecordResult;
  exitDigit: number;
}

export type EngineStatusState = typeof EngineStatusState[keyof typeof EngineStatusState];


export const EngineStatusState = {
  idle: 'idle',
  running: 'running',
  paused: 'paused',
  stopped: 'stopped',
} as const;

export interface EngineStatus {
  state: EngineStatusState;
  config: EngineConfig;
  /** @nullable */
  accountId: string | null;
  currentStake: number;
  successiveLosses: number;
  successiveWins: number;
  sessionPnl: number;
  totalTrades: number;
  wins: number;
  losses: number;
  /** @nullable */
  lastTick: number | null;
  /** @nullable */
  lastDigit: number | null;
  windowFillCount: number;
  windowSize: number;
  /** @nullable */
  under9Confidence: number | null;
  /** @nullable */
  under8Confidence: number | null;
  /** @nullable */
  stopReason: string | null;
  /** @nullable */
  errorMessage: string | null;
  /** @nullable */
  startedAt: string | null;
  recentTrades: TradeRecord[];
}