# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /engine/config:
    patch:
      operationId: updateEngineConfig
      tags: [engine]
      summary: Update engine configuration
      description: Updates one or more configuration fields. The engine starts fully idle by default and configuration can be changed at any time; changing market or accountType while running requires pausing or stopping first.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EngineConfigUpdate"
      responses:
        "200":
          description: Updated engine status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngineStatus"
        "400":
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Cannot change this field while the engine is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /engine/start:
    post:
      operationId: startEngine
      tags: [engine]
      summary: Start (or resume) the trading engine
      description: Connects to Deriv using the account type in the current config, subscribes to ticks, and begins evaluating the Under 9 / Under 8 digit strategy. Resuming from paused continues the same session; starting from idle or stopped begins a fresh session and resets session stats.
      responses:
        "200":
          description: Engine status after starting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngineStatus"
        "400":
          description: Cannot start (e.g. missing Deriv token for selected account type)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    EngineConfig:
      type: object
      required:
        - market
        - accountType
        - baseStake
        - martingaleEnabled
        - martingaleMultiplier
        - stopLoss
        - takeProfit
        - maxSuccessiveLosses
        - maxSuccessiveWins
      properties:
        market:
          type: string
          description: Deriv volatility index symbol, e.g. R_100
        accountType:
          type: string
          enum: [demo, real]
        baseStake:
          type: number
          minimum: 0.35
        martingaleEnabled:
          type: boolean
        martingaleMultiplier:
          type: number
          minimum: 1
        stopLoss:
          type: number
          minimum: 0
        takeProfit:
          type: number
          minimum: 0
        maxSuccessiveLosses:
          type: integer
          minimum: 1
        maxSuccessiveWins:
          type: integer
          minimum: 1
    EngineConfigUpdate:
      type: object
      properties:
        market:
          type: string
        accountType:
          type: string
          enum: [demo, real]
        baseStake:
          type: number
          minimum: 0.35
        martingaleEnabled:
          type: boolean
        martingaleMultiplier:
          type: number
          minimum: 1
        stopLoss:
          type: number
          minimum: 0
        takeProfit:
          type: number
          minimum: 0
        maxSuccessiveLosses:
          type: integer
          minimum: 1
        maxSuccessiveWins:
          type: integer
          minimum: 1
    TradeRecord:
      type: object
      required:
        - id
        - timestamp
        - market
        - contractType
        - barrier
        - stake
        - payout
        - profit
        - result
        - exitDigit
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        market:
          type: string
        contractType:
          type: string
        barrier:
          type: string
        stake:
          type: number
        payout:
          type: number
        profit:
          type: number
        result:
          type: string
          enum: [win, loss]
        exitDigit:
          type: integer
    EngineStatus:
      type: object
      required:
        - state
        - config
        - accountId
        - currentStake
        - successiveLosses
        - successiveWins
        - sessionPnl
        - totalTrades
        - wins
        - losses
        - lastTick
        - lastDigit
        - windowFillCount
        - windowSize
        - under9Confidence
        - under8Confidence
        - stopReason
        - errorMessage
        - startedAt
        - recentTrades
      properties:
        state:
          type: string
          enum: [idle, running, paused, stopped]
        config:
          $ref: "#/components/schemas/EngineConfig"
        accountId:
          type: ["string", "null"]
        currentStake:
          type: number
        successiveLosses:
          type: integer
        successiveWins:
          type: integer
        sessionPnl:
          type: number
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        lastTick:
          type: ["number", "null"]
        lastDigit:
          type: ["integer", "null"]
        windowFillCount:
          type: integer
        windowSize:
          type: integer
        under9Confidence:
          type: ["number", "null"]
        under8Confidence:
          type: ["number", "null"]
        stopReason:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
          format: date-time
        recentTrades:
          type: array
          items:
            $ref: "#/components/schemas/TradeRecord"