# lib/api-zod/src/generated/types/engineConfigUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EngineConfigUpdateAccountType } from './engineConfigUpdateAccountType';

export interface EngineConfigUpdate {
  market?: string;
  accountType?: EngineConfigUpdateAccountType;
  /** @minimum 0.35 */
  baseStake?: number;
  martingaleEnabled?: boolean;
  /** @minimum 1 */
  martingaleMultiplier?: number;
  /** @minimum 0 */
  stopLoss?: number;
  /** @minimum 0 */
  takeProfit?: number;
  /** @minimum 1 */
  maxSuccessiveLosses?: number;
  /** @minimum 1 */
  maxSuccessiveWins?: number;
}