# lib/api-zod/src/generated/types/engineStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EngineConfig } from './engineConfig';
import type { EngineStatusState } from './engineStatusState';
import type { TradeRecord } from './tradeRecord';

export interface EngineStatus {
  state: EngineStatusState;
  config: EngineConfig;
  /** @nullable */
  accountId: string | null;
  currentStake: number;
  successiveLosses: number;
  successiveWins: number;
  sessionPnl: number;
  totalTrades: number;
  wins: number;
  losses: number;
  /** @nullable */
  lastTick: number | null;
  /** @nullable */
  lastDigit: number | null;
  windowFillCount: number;
  windowSize: number;
  /** @nullable */
  under9Confidence: number | null;
  /** @nullable */
  under8Confidence: number | null;
  /** @nullable */
  stopReason: string | null;
  /** @nullable */
  errorMessage: string | null;
  /** @nullable */
  startedAt: Date | null;
  recentTrades: TradeRecord[];
}