# lib/api-zod/src/generated/types/tradeRecord.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeRecordResult } from './tradeRecordResult';

export interface TradeRecord {
  id: string;
  timestamp: Date;
  market: string;
  contractType: string;
  barrier: string;
  stake: number;
  payout: number;
  profit: number;
  result: TradeRecordResult;
  exitDigit: number;
}